requirements.txt
# Core Django
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-storages==1.14.6
python-decouple==3.8

# Database
psycopg2-binary==2.9.11
mysqlclient==2.2.7

# Core dependencies
asgiref==3.11.0
sqlparse==0.5.5
PyJWT==2.10.1
boto3==1.41.5

# Face recognition stack
face-recognition==1.3.0
face_recognition_models==0.3.0
opencv-python-headless==4.10.0.84

# Scientific computing
numpy==2.2.6
scipy==1.16.3
scikit-image==0.26.0
pillow==12.0.0

# Additional dependencies
ImageIO==2.37.2
lazy_loader==0.4
networkx==3.6.1
packaging==25.0
tifffile==2025.12.20
openpyxl==3.1.5
reportlab==4.4.7
click==8.3.1
django-apscheduler==0.7.0
charset-normalizer==3.4.4
et_xmlfile==2.0.0

# Production server
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==3.0.1